[M]支持各版本游戏下载对应的首包资源

+++ Unity/Script/Script_Main/ResDownload_multiBatch.cs
@@ -8,6 +8,15 @@ using Shark;
 using UnityEngine.Networking;
 using ICSharpCode.SharpZipLib.Zip;
 
+public class Route{
+    public string appVer;
+    public string baseVer;
+    public string route;
+}
+public class PatchPath
+{
+    public Route[] routes;
+}
 public class Patch {
     public string name;
     public string md5;
@@ -20,15 +29,21 @@ public class PatchCfg {
 partial class ResDownload
 {
     PatchCfg m_PatchCfg;
+    PatchPath m_PatchPath;
     public long TotalPatchSize = 0;
     private bool HaveTipSpace = false;
     IEnumerator CheckAndDownloadPatch()
     {
         yield return m_UILogic.StartCoroutine(DownPatchCfg());
+        if (null == m_PatchCfg) {
+            Debug.LogError("无基础资源包");
+            yield break;
+        }
         List<Patch> dlLst = new List<Patch>();
         yield return m_UILogic.StartCoroutine(CalAndGetDownloadPatch(dlLst));
         if (!HaveTipSpace && TotalPatchSize > 0)
         {
+            m_UILogic.SetProgressTips(UpdateTips.GetTipByID(203));
             var totalLength = TotalPatchSize;
             string strTips = string.Format(UpdateTips.GetTipByID(33), totalLength * 1f / 1024 / 1024);
 
@@ -71,18 +86,66 @@ partial class ResDownload
     /// </summary>
     /// <returns></returns>
     public IEnumerator DownPatchCfg() {
-        if (null != m_PatchCfg) {
+        if (null != m_PatchCfg && null != m_PatchPath) {
             yield break;
         }
-        string firstUrl = URLConfig.Instance.GetURLCDN(DownloadHelper.AddTimestampToUrl(PathConfig.Instance.MultiPatchCfg));
-        var request = UnityWebRequest.Get(firstUrl);
+        #region 获取补丁路由文件
+        string patchPathUrl = URLConfig.Instance.GetPatchPathURL(DownloadHelper.AddTimestampToUrl(PathConfig.PatchPath));
+        if (EnvUtils.IsDEVELOPMENT_BUILD() || EnvUtils.IsUnity_Editor())
+        {
+            Debug.Log($"### patchPathUrl:{patchPathUrl}");
+        }
+        var request = UnityWebRequest.Get(patchPathUrl);
+        yield return request.SendWebRequest();
+        string patchPathTxt = request.downloadHandler.text;
+        if (string.IsNullOrEmpty(request.error))
+        {
+            if (EnvUtils.IsDEVELOPMENT_BUILD() || EnvUtils.IsUnity_Editor())
+            {
+                Debug.Log($"PatchPath: {patchPathTxt}");
+            }
+            try
+            {
+                JsonData jd = JsonMapper.ToObject(patchPathTxt);
+                if (null != jd)
+                {
+                    m_PatchPath = JsonMapper.ToObject<PatchPath>(jd.ToJson());
+                }
+            }
+            catch (Exception e)
+            {
+                Debug.LogError(e.ToString());
+            }
+        }
+        else
+        {
+            Debug.LogError($"### {request.error} {patchPathUrl}");
+        }
+        #endregion
+        if (null == m_PatchPath) {
+            yield break;
+        }
+        string patchRoute = "";
+        foreach (var route in m_PatchPath.routes) {
+            if (route.appVer == Application.version.Trim() && route.baseVer == m_stringAppVersion) {
+                patchRoute = route.route;
+                break;
+            }
+        }
+        if (string.IsNullOrEmpty(patchRoute)) {
+            Debug.LogError($"### 通知研发人员配置首包资源路径 version:{Application.version} baseRes:{m_stringAppVersion}");
+            yield break;
+        }
+        PathConfig.PatchRoute = patchRoute;
+        string patchCfgURl = URLConfig.Instance.GetPatchURL(DownloadHelper.AddTimestampToUrl(PathConfig.Instance.MultiPatchCfg));
+        request = UnityWebRequest.Get(patchCfgURl);
         yield return request.SendWebRequest();
         string patchCfgTxt = request.downloadHandler.text;
         if (string.IsNullOrEmpty(request.error))
         {
             if (EnvUtils.IsDEVELOPMENT_BUILD() || EnvUtils.IsUnity_Editor())
             {
-                Debug.Log($"DownBatchCfg: {patchCfgTxt}");
+                Debug.Log($"### PatchCfg: {patchCfgTxt}");
             }
             try
             {
@@ -99,7 +162,10 @@ partial class ResDownload
         }
         else
         {
-            Debug.LogError($"### {request.error}");
+            Debug.LogError($"### {request.error} {patchCfgURl}");
+        }
+        if (null == m_PatchCfg) {
+            Debug.LogError($"### 补丁配置获取失败，请获取日志并发给研发人员");
         }
     }
     /// <summary>
@@ -107,17 +173,19 @@ partial class ResDownload
     /// </summary>
     /// <param name="lst"></param>
     IEnumerator CalAndGetDownloadPatch(List<Patch> lst) {
+        m_UILogic.SetProgressTips(UpdateTips.GetTipByID(202));
         TotalPatchSize = 0;
         for (int index = 0; index < m_PatchCfg.patchs.Length; index++)
         {
             m_UILogic.SetProgress(index * 1f / m_PatchCfg.patchs.Length);
+            bool needDownload = true;
             var patch = m_PatchCfg.patchs[index];
             var cachePatch = PathConfig.Instance.DownLoadCachePath + "/" + patch.name;
             if (File.Exists(cachePatch))
             {
                 if (GfxUtils.GetMD5Hash(cachePatch).Equals(patch.md5))
                 {
-                    continue;
+                    needDownload = false;
                 }
                 else
                 {
@@ -125,25 +193,31 @@ partial class ResDownload
                 }
             }
             yield return new WaitForEndOfFrame();
-            if (long.TryParse(patch.size, out var size))
+            if (needDownload)
             {
-                TotalPatchSize += size;
+                if (long.TryParse(patch.size, out var size))
+                {
+                    TotalPatchSize += size;
+                }
+                else
+                {
+                    Debug.LogError($"{patch.name} Size[{patch.size}] invaild.");
+                }
+                lst.Add(patch);
             }
-            else 
-            {
-                Debug.LogError($"{patch.name} Size[{patch.size}] invaild.");
-            }
-            lst.Add(patch);
         }
+        this.SetProgress(1f);
     }
     /// <summary>
     /// 开始下载补丁包
     /// </summary>
     IEnumerator DownMultiBatch(List<Patch> lst) {
+        ChanConnector.SendGameInfoLog("basezip_donwload", "basezip_donwload", "basezip_donwload_confirm_download", "basezip_donwload_confirm_download");
         m_UILogic.SetProgressTips(UpdateTips.GetTipByID(200));
         m_curUpdateStep = UpdateStep.DownBaseRes;
         bool downloadPatchFail = false;
         string errorTip = string.Empty;
+        lastReportProgress = 0;
         yield return new HttpDownLoad().Start(
                 this.m_UILogic,
                 lst,
@@ -162,6 +236,11 @@ partial class ResDownload
                     }
                 },
                 (f)=> {
+                    if ((f - lastReportProgress) * 100 >= 20)
+                    {
+                        ChanConnector.SendGameInfoLog("basezip_donwload", "basezip_donwload", $"basezip_donwload_progress_{Mathf.RoundToInt(f * 100)}", $"basezip_donwload_progress_{Mathf.RoundToInt(f * 100)}", $"progress:{Mathf.RoundToInt(f * 100)}");
+                        lastReportProgress = f;
+                    }
                     m_UILogic.SetProgress(f);
                 },
                 (downloadedSize, allSize, sizePerSec) => {
@@ -181,22 +260,28 @@ partial class ResDownload
         yield return new WaitForEndOfFrame();
         if (downloadPatchFail)
         {
+            ChanConnector.SendGameInfoLog("basezip_donwload", "basezip_donwload", "basezip_donwload_fail", "basezip_donwload_fail", $"fail {errorTip}");
             m_UILogic.ShowMessageBox(errorTip, () => { 
                 m_UILogic.StartCoroutine(CheckAndDownloadPatch()); 
             });
         }
         else
         {
+            ChanConnector.SendGameInfoLog("basezip_donwload", "basezip_donwload", "basezip_donwload_success", "basezip_donwload_success", "success");
             m_UILogic.StartCoroutine(CheckAndDownloadPatch());
         }
+        ChanConnector.SendGameInfoLog("basezip_donwload", "basezip_donwload", "basezip_donwload_end", "basezip_donwload_end");
     }
     IEnumerator UnzipPatch(PatchCfg patchCfg) {
         List<Patch> lst = new List<Patch>();
         foreach (var patch in patchCfg.patchs) {
             lst.Add(patch);
         }
+        lastReportProgress = 0;
         m_UILogic.SetProgressTips(UpdateTips.GetTipByID(300));
         yield return null;
+        ChanConnector.SendGameInfoLog("basezip_uncompress", "basezip_uncompress", "basezip_uncompress_begin", "basezip_uncompress_begin");
+        string errorTip = string.Empty;
         yield return m_UILogic.StartCoroutine(MyZipTool.UnZipFileCo(lst, PathConfig.Instance.PersistenPath, OnProgressChange: (f, result, msg) =>
         {
             if (CurUnZipStep != result)
@@ -210,40 +295,50 @@ partial class ResDownload
             else if (result == MyZipTool.UnZipStep.Progress)
             {
                 this.m_UILogic.SetProgressTxt(f);
+                if ((f - lastReportProgress) * 100 >= 20)
+                {
+                    ChanConnector.SendGameInfoLog("basezip_uncompress", "basezip_uncompress", $"basezip_uncompress_progress_{Mathf.RoundToInt(f * 100)}", $"basezip_uncompress_progress_{Mathf.RoundToInt(f * 100)}", $"progress:{Mathf.RoundToInt(f * 100)}");
+                    lastReportProgress = f;
+                }
             }
             else if (result == MyZipTool.UnZipStep.Fail)
             {
-                errorMsg = msg;
-                if (errorMsg.Contains("Disk full"))
-                {
-                    m_UILogic.SetProgress(1f);
-                    m_UILogic.SetProgressTips(UpdateTips.GetTipByID(304));
-                    m_UILogic.ShowMessageBox(UpdateTips.GetTipByID(304), () =>
-                    {
-                    });
-                }
-                else
-                {
-                    m_UILogic.SetProgress(1f);
-                    m_UILogic.SetProgressTips(UpdateTips.GetTipByID(301));
-                    m_UILogic.ShowMessageBox(UpdateTips.GetTipByID(302), () =>
-                    {
-                        CheckUpdateBasePatch();
-                    });
-                }
+                errorTip = msg;
             }
         }, OnUnzipSpdChange: (sizePerSec) =>
         {
-            string errorTip = string.Format(UpdateTips.GetTipByID(303), sizePerSec * 1f / 1024 / 1024);
-            m_UILogic.SetProgressTips3(errorTip);
+            m_UILogic.SetProgressTips3(string.Format(UpdateTips.GetTipByID(303), sizePerSec * 1f / 1024 / 1024));
         }));
 
-        File.Move(PathConfig.Instance.PersistenPath + PathConfig.Instance.VersionFileName, PathConfig.Instance.LocalVersionFilePath);
-        if (Directory.Exists(PathConfig.Instance.DownLoadCachePath))
+        if (string.IsNullOrEmpty(errorTip))
         {
-            Directory.Delete(PathConfig.Instance.DownLoadCachePath, true);
+            ChanConnector.SendGameInfoLog("basezip_uncompress", "basezip_uncompress", "basezip_uncompress_success", "basezip_uncompress_success");
+            File.Move(PathConfig.Instance.PersistenPath + PathConfig.Instance.VersionFileName, PathConfig.Instance.LocalVersionFilePath);
+            if (Directory.Exists(PathConfig.Instance.DownLoadCachePath))
+            {
+                Directory.Delete(PathConfig.Instance.DownLoadCachePath, true);
+            }
+            this.SetProgress(1f);
+            ReCheckResVersion();
         }
-        this.SetProgress(1f);
-        ReCheckResVersion();
+        else
+        {
+            ChanConnector.SendGameInfoLog("basezip_uncompress", "basezip_uncompress", "basezip_uncompress_fail", "basezip_uncompress_fail", $"fail:{errorTip}");
+            m_UILogic.SetProgress(1f);
+            if (errorTip.Contains("Disk full"))
+            {
+                m_UILogic.SetProgressTips(UpdateTips.GetTipByID(304));
+            }
+            else
+            {
+                m_UILogic.SetProgressTips(UpdateTips.GetTipByID(301));
+            }
+            m_UILogic.ShowMessageBox(UpdateTips.GetTipByID(302), () =>
+            {
+                m_UILogic.StartCoroutine(CheckAndDownloadPatch());
+            });
+        }
+        ChanConnector.SendGameInfoLog("basezip_uncompress", "basezip_uncompress", "basezip_uncompress_end", "basezip_uncompress_end");
+        
     }
 }
